refactor(diff): split pkg/diff into cohesive files
The 415-line diff.go had grown into a god-file mixing format taxonomy,
pairing, normalization, and rendering. Split it into format.go, pair.go,
normalize.go, and render.go (with mirrored test files), rewrite the
stale package doc to cover all the new styles, and polish the render
internals (a shared writeBody, a typed changeKind). No behavior change.
